Bump Block & Transaction version and share with ZNP
Currently there is a single deployed version 0.
With this change we will continue to have a signle version 1 for every thing.
Unless there is a compeling reason we probably want to continue this way for now.

This version should be incremented after every deployed release.

Test:
cargo test
cargo clippy

Use in ZNP/tests & clippy